refactor - change shop page folder structure
