+++ components/Shop/Organisms/ShopHeader/style.module.scss
@@ -1,23 +1,9 @@
 .header {
-  display: flex;
-  @include minMobile() {
-    flex-wrap: wrap;
-  }
-  align-items: center;
-  padding: 13px 23px;
-  .logo {
-    text-align: left;
-  }
-  &-tool {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    .btn {
-      display: flex;
-      &-search {
-        display: flex;
-        margin: 0 20px 0 auto;
-      }
-    }
-  }
+  z-index: 1;
+  position: fixed;
+  top: 0;
+  max-width: 1024px;
+  width: 100%;
+  padding-top: 15px;
+  background-color: $white;
 }
