+++ components/Shop/molecules/HeaderTool/style.module.scss
@@ -1,0 +1,23 @@
+.tool-container {
+  display: flex;
+  @include minMobile() {
+    flex-wrap: wrap;
+  }
+  align-items: center;
+  padding: 13px 23px;
+  .logo {
+    text-align: left;
+  }
+  .header-tool {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    .btn {
+      display: flex;
+      &-search {
+        display: flex;
+        margin: 0 20px 0 auto;
+      }
+    }
+  }
+}
